feat(recipe-import): YouTube video extraction + gemini-service refactor
Adds a third tab to the recipe import modal that ingests YouTube watch URLs,
Shorts, and youtu.be links and extracts a structured recipe via Gemini's
fileData video input. Reuses the existing RECIPE_SCHEMA for validation.

To keep functions/ maintainable as it grows, splits the 657-line
gemini-service.js into purpose-based modules under functions/utils/gemini/
(extractors/{image,url,youtube}, enhancement/food-image, client, recipe-schema)
with a thin barrel re-export. Image enhancement intentionally sits in a
sibling folder, not under extractors/, because it produces images, not
recipe data.

Also extracts the auth+role gate from functions/index.js into
withAuthAndRole() so the new video callable doesn't repeat the pattern
a fourth time.

Frontend:
- New 'video' import mode alongside 'image' and 'url' in recipe_import_modal.
- Shared YouTube URL parser at src/js/utils/youtube-url.js with 20 unit
  tests at 100% coverage (valid watch/Shorts/youtu.be/embed/mobile/music
  URLs; rejects vimeo, channel pages, malformed IDs, non-strings).
- Mode-aware error messages cover the video failure modes
  (invalid YouTube URL, private/age-restricted, no recipe in video).

Backend:
- New extractRecipeFromVideo callable, role-gated to approved|manager,
  with 300s timeout for longer video processing.
- Failures land in failed_url_extractions with kind: 'video' (existing
  url-extraction failures now also tag kind: 'url').
- Bad YouTube URLs surface as invalid-argument so the client can show a
  precise message; everything else maps to internal.
